- moved TaskList class definition to its own file - update colors used in activity indicators
